+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 * Feature: Merged [pull request 211](https://github.com/rubiii/savon/pull/211) by
   [mattkirman](https://github.com/mattkirman) to fix [issue 202](https://github.com/rubiii/savon/issues/202).
 
+* Fix: [issue 218](https://github.com/rubiii/savon/pull/218) - Savon now correctly handles namespaced
+  Array items in a Hash passed to `Savon::SOAP::XML#body=`.
+
 ## 0.9.6 (2011-07-07)
 
 * Improvement/Fix: Updated Savon to use the latest version of [Wasabi](http://rubygems.org/gems/wasabi).