+++ CHANGELOG.md
@@ -22,6 +22,9 @@
 * Fix: [issue 218](https://github.com/rubiii/savon/pull/218) - Savon now correctly handles namespaced
   Array items in a Hash passed to `Savon::SOAP::XML#body=`.
 
+* Fix: Merged [pull request 212](https://github.com/rubiii/savon/pull/212) to fix
+  [savon_spec issue 2](https://github.com/rubiii/savon_spec/issues/2).
+
 * Improvement: [issue 222](https://github.com/rubiii/savon/issues/222) - Set the Content-Length header.
 
 ## 0.9.6 (2011-07-07)